requirements.txt
# Cloud sync dependency libs
boto>=2.29.1
django>=1.6.5
django-storages>=1.1
requests>=2.3.0
PyYAML>=3.11

# Auto deploy tool
fabric >=1.5
